+++ static/events/monster_still_fight.txt
